impvoements planner and model with agent embedding

+++ testing_agents/test_single_agent_v2.py
@@ -1,4 +1,5 @@
 import sys
+from termcolor import colored
 import shutil
 import os
 import logging
@@ -38,8 +39,8 @@ if __name__ == '__main__':
     args.max_episode_length = 150
     args.num_per_apartment = 20
     
-    #args.dataset_path = './dataset/test_env_task_set_10_full_reduced_tasks_single.pik'
-    args.dataset_path = './dataset/train_env_task_set_20_full_reduced_tasks1to3.pik'
+    args.dataset_path = './dataset/test_env_task_set_10_full_reduced_tasks1to3.pik'
+    #args.dataset_path = './dataset/train_env_task_set_20_full_reduced_tasks1to3.pik'
 
     # Beliefs
     # spiked: object is in cabinet
@@ -58,19 +59,16 @@ if __name__ == '__main__':
             ['partial', 0, 0.05, False, 0, "spiked2"], # 10 High prior for not inside
             ['partial', 0, 0.05, False, 0, "spiked3"], # 11 For sure not in bathroom
             ['partial', 0, 0.05, False, 0, "spiked4"], # 12 All things kithcen
-            ['partial', 0, 0.05, False, 0.1, "spiked"], # 14
-            ['partial', 0, 0.05, False, 0.1, "spiked2"] # 15
+            ['partial', 0, 0.05, False, 0.1, "spiked"], # 13
+            ['partial', 0, 0.05, False, 0.1, "spiked2"] # 14
     ]
     random_start = random.Random()
     agent_types_index = list(range(9))
-    agent_types_index =  [0, 3, 4, 10, 12, 13, 14, 15]
-    agent_types_index = [10]
-    #random.shuffle(agent_types_index)
+    agent_types_index =  [0, 3, 4, 10, 12, 13, 14]
+    random_start.shuffle(agent_types_index)
     if args.agenttype != 'all':
         agent_types_index = [int(x) for x in args.agenttype.split(',')]
     for agent_id in agent_types_index: #len(agent_types)):
-        if agent_id in [4]:
-            continue
         args.obs_type, open_cost, walk_cost, should_close, forget_rate, belief_type = agent_types[agent_id]
         datafile = args.dataset_path.split('/')[-1].replace('.pik', '')
         agent_args = {
@@ -81,13 +79,18 @@ if __name__ == '__main__':
             'belief': {'forget_rate': forget_rate, 'belief_type': belief_type}
         }
         args.mode = '{}_'.format(agent_id+1) + get_class_mode(agent_args)
-        args.mode += 'v9_particles_v2'
+        args.mode += 'v9_particles_v2_modeinfo'
 
         
         env_task_set = pickle.load(open(args.dataset_path, 'rb'))
         print(len(env_task_set))
 
         for env in env_task_set:
+            # Remove one of the goals
+            for goal_pred in env['task_goal'][0]:
+                if 'sit' in goal_pred:
+                    env['task_goal'][0][goal_pred] = 0
+
             init_gr = env['init_graph']
             gbg_can = [node['id'] for node in init_gr['nodes'] if node['class_name'] in ['garbagecan', 'clothespile']]
             init_gr['nodes'] = [node for node in init_gr['nodes'] if node['id'] not in gbg_can]
@@ -96,8 +99,8 @@ if __name__ == '__main__':
                 if node['class_name'] == 'cutleryfork':
                     node['obj_transform']['position'][1] += 0.1
 
-        args.record_dir = '../data_scratch/large_data/{}/{}'.format(datafile, args.mode)
-        error_dir = '../data_scratch/large_data/logging/{}_{}'.format(datafile, args.mode)
+        args.record_dir = '../data_scratch/large_data_v2/{}/{}'.format(datafile, args.mode)
+        error_dir = '../data_scratch/large_data_v2/logging/{}_{}'.format(datafile, args.mode)
         if not os.path.exists(args.record_dir):
             os.makedirs(args.record_dir)
 
@@ -106,7 +109,7 @@ if __name__ == '__main__':
 
         executable_args = {
                         'file_name': args.executable_file,
-                        'x_display': args.display,
+                        'x_display': args.display if args.saveimg else None,
                         'no_graphics': not args.saveimg
         }
 
@@ -122,8 +125,9 @@ if __name__ == '__main__':
         
         test_results = {}
         #episode_ids = [episode_ids[0]]
-        episode_ids = [185]
-     
+        #episode_ids = [185]
+        
+        file_failures = 'failures_{}.txt'.format(args.base_port)
         def env_fn(env_id):
             return UnityEnvironment(num_agents=1,
                                     max_episode_length=args.max_episode_length,
@@ -154,7 +158,7 @@ if __name__ == '__main__':
         arena = ArenaMP(args.max_episode_length, id_run, env_fn, agents)
         
         # episode_ids = [20] #episode_ids
-        # num_tries = 1
+        num_tries = 2
         episode_ids = episode_ids
 
         for iter_id in range(num_tries):
@@ -184,8 +188,10 @@ if __name__ == '__main__':
                 log_file_name = args.record_dir + '/logs_episode.{}_iter.{}.pik'.format(episode_id, iter_id)
                 failure_file = '{}/{}_{}.txt'.format(error_dir, episode_id, iter_id)
 
+                if sum(env_task_set[episode_id]['task_goal'][0].values()) == 0:
+                    continue
                 if os.path.isfile(log_file_name):# or os.path.isfile(failure_file):
-                    #pass
+                    # pass
                     continue
                 if os.path.isfile(failure_file):
                     os.remove(failure_file)
@@ -199,7 +205,7 @@ if __name__ == '__main__':
                 for it_agent, agent in enumerate(arena.agents):
                     agent.seed = (it_agent + current_tried * 2) * 5
 
-                
+                failure = False
                 try:
                     
                     arena.reset(episode_id)
@@ -222,6 +228,7 @@ if __name__ == '__main__':
 
                     Path(args.record_dir).mkdir(parents=True, exist_ok=True)
                     if len(saved_info['obs']) > 0:
+                        print(colored("Saving.."), 'green')
                         pickle.dump(saved_info, open(log_file_name, 'wb'))
                     else:
                         with open(log_file_name, 'w+') as f:
@@ -236,7 +243,8 @@ if __name__ == '__main__':
                     print("Unity exception")
                     arena.reset_env()
                     #ipdb.set_trace()
-                    continue
+                    failure = True
+                    failure_str = 'unity'
 
                 except utils_exception.ManyFailureException as e:
                     traceback.print_exc()
@@ -248,8 +256,19 @@ if __name__ == '__main__':
                     #exit()
                     #arena.reset_env()
                     print("Dione")
-                    ipdb.set_trace()
+
+                    #Path(args.record_dir).mkdir(parents=True, exist_ok=True)
+                    saved_info = arena.saved_info
+                    failure = True
+                    #print("ACTIONS")
+                    #for action in arena.saved_info['action'][0]:
+                    #    print(action)
+                    #print('---')
+                    #if len(saved_info['obs']) > 0:
+                    #    print(colored("Saving.."), 'green')
+                    #    pickle.dump(saved_info, open(log_file_name, 'wb'))
                     arena.reset_env()
+                    failure_str = "many_actions"
                     continue
 
                 except Exception as e:
@@ -262,11 +281,12 @@ if __name__ == '__main__':
                     #    f.write(error_str)
                     traceback.print_exc()
 
-                    Path(args.record_dir).mkdir(parents=True, exist_ok=True)
-                    saved_info = arena.saved_info
-                    if len(saved_info['obs']) > 0:
-                        pickle.dump(saved_info, open(log_file_name, 'wb'))
-                    ipdb.set_trace()
+                    #Path(args.record_dir).mkdir(parents=True, exist_ok=True)
+                    #saved_info = arena.saved_info
+                    #if len(saved_info['obs']) > 0:
+                    #    print(colored("Saving.."), 'green')
+                    #    pickle.dump(saved_info, open(log_file_name, 'wb'))
+                    #ipdb.set_trace()
 
                     logging.exception("Error")
                     print("OTHER ERROR")
@@ -276,13 +296,22 @@ if __name__ == '__main__':
                     # ipdb.set_trace()
                     # ipdb.set_trace()
                     # pdb.set_trace()
+                    failure = True
+                    failure_str = "other"
                     continue
-                S[episode_id].append(is_finished)
-                L[episode_id].append(steps)
-                test_results[episode_id] = {'S': S[episode_id],
-                                            'L': L[episode_id]}
+
+                if failure:
+                    with open(file_failures, 'a+'):
+                        str_file = 'Episode: {}. Try: {}. Agent: {}. Failure: {}\n'.format(episode_id, iter_id, agent_id, failure_str)
+                        f.write(str_file)
+
+
+                #S[episode_id].append(is_finished)
+                #L[episode_id].append(steps)
+                #test_results[episode_id] = {'S': S[episode_id],
+                #                            'L': L[episode_id]}
                                             
             print('average steps (finishing the tasks):', np.array(steps_list).mean() if len(steps_list) > 0 else None)
             print('failed_tasks:', failed_tasks)
-            pickle.dump(test_results, open(args.record_dir + '/results_{}.pik'.format(0), 'wb'))
+            #pickle.dump(test_results, open(args.record_dir + '/results_{}.pik'.format(0), 'wb'))
